Link to JDK Supplier, just as we link to JDK Function and Predicate.
Change Function and Predicate to display the fully qualified name of the class they're linking to.

Also link to the JDK versions of all those classes in the backport, which we're building with a Java 8 version of Javadoc and from which we're linking to Java 9 Javadoc.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=178673695