Make lenientToString easier for nullness-analysis tools to understand.
And I think prevent it from turning a StackOverflowError into a NullPointerException in unlikely cases?

As far as tools know, String.valueOf is capable of throwing an Error or RuntimeException, even when it's called with a parameter whose values is null. Thus, tools conclude that the flow of the program might reach `o.getClass().getName()` when `o` is null. That would in turn produce NullPointerException.

Tools might not even be wrong, since String.valueOf could produce a StackOverflowError if we were *just* about to run out of space. Of course, in that case we're going to have problems in the catch() block no matter what... :)

To avoid this, we check for null in lenientToString itself.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=297360806