Ensure QueueWorker does not outlive SequentialExecutor on iOS.
When `SequentialExecutor` is used on iOS as J2ObjC-transpiled code, we need to
make sure that `SequentialExecutor.QueueWorker` does not outlive its outer
class. Previously, because the inner class was annotated with `@WeakOuter`, it
could happen, and, when this happened, the inner class would then invoke
methods on a dangling pointer and crash the process in turn. The issue is now
fixed by annotating the worker field in the outer class with `@RetainedWith`.

Although `@RetainedWith` was initially designed for bimaps, the annotation
addresses exactly the lifecycle problem here.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=311046012