Fix comment in Converter.
Converter.apply is annotated as returning null, so it's sound. The LegacyConverter problem arises only if users view the converter as a Function<A, B>, since Function<A, B>.apply is *not* annotated as returning null (and yet a LegacyConverter may do so, even given a non-null input).

RELNOTES=n/a
PiperOrigin-RevId: 368219982