Remove `@Beta` from `AsyncCallable`, `ListenableScheduledFuture`, and `ClosingFuture`.
The former two are simple interfaces. `ClosingFuture` was meant to be `@Beta` for only one release :)

`ListenableScheduledFuture` was particularly silly to have as `@Beta` because it's the return type of `ListeningScheduledExecutorService`, from which we'd already removed `@Beta`.

`AsyncCallable` makes extra sense to remove from `@Beta` at the same time as `Futures.submit`, `submitAsync`, and `scheduleAsync` (CL 416093023).

RELNOTES=`util.concurrent`: Removed `@Beta` from `AsyncCallable`, `ListenableScheduledFuture`, and `ClosingFuture`.
PiperOrigin-RevId: 416302588