Address an Error Prone warning about catching `AssertionError`.
Error Prone was warning that if an assertion failed, the `AssertionError` would be caught. It turns out that the exception would in fact have been propagated and rethrown to fail the test, but to see that required close inspection, and the code may also not have been thread-safe. It is much more straightforward just to use an `Executor`, whereby exceptions get propogated into `Future.get`.

RELNOTES=n/a
PiperOrigin-RevId: 563164302