Annotate hash-flooding tests as `@GwtIncompatible`.
This replaces our old "strategy" of annotating some of them as `@GwtCompatible` but then annotating both their test methods as `@GwtIncompatible`. That "strategy" results in a number of empty test classes.

Those empty tests classes don't upset GWT (maybe because of cl/44844299?), but they do upset J2CL. (And they're slightly wasteful and confusing, too.)

As a result of this, I can also roll back a small part of cl/570444724, which had adapted one of the hash-flooding tests to work under J2CL.

PiperOrigin-RevId: 570489131