Avoid exposing some types outside the scope in which those types are visible.
Java doesn't mind this, but it's not like it's accomplishing anything for us.

I've tried to make things `private` where possible and fallen back to making types package-private where necessary.

(In `Synchronized`, I made everything package-private for consistency, rather than trying to trace through the class hierarchies to expose only the types that need it.)

I also sprinkled in some `final` keywords while I was in the area.

RELNOTES=n/a
PiperOrigin-RevId: 573211489